+++ local_test_settings.py
@@ -2,5 +2,4 @@ from .base import *
 
 INSTALLED_APPS += (
     'wagtailsurveys',
-    'molo.surveys'
 )