+++ molo/surveys/models.py
@@ -5,8 +5,9 @@ from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
 from django.db.models.fields import TextField, BooleanField
-from django.shortcuts import render
-from django.dispatch import receiver
+from django.shortcuts import render, redirect
+from django.core.urlresolvers import reverse
+
 from modelcluster.fields import ParentalKey
 
 from molo.core.models import (
@@ -53,6 +54,7 @@ def create_survey_index_page(sender, instance, **kwargs):
 
 class MoloSurveyPage(
         TranslatablePageMixinNotRoutable, surveys_models.AbstractSurvey):
+
     intro = TextField(blank=True)
     thank_you_text = TextField(blank=True)
 
@@ -94,50 +96,6 @@ class MoloSurveyPage(
             FieldPanel('multi_step')
         ], heading='Survey Settings')
     ]
-
-    def get_context(self, request, *args, **kwargs):
-        context = super(MoloSurveyPage, self).get_context(
-            request, *args, **kwargs
-        )
-
-        # check request method so that results are shown only on the landing
-        # page
-        if self.show_results and request.method == 'POST':
-            results = dict()
-            # Get information about form fields
-            data_fields = [
-                (field.clean_name, field.label)
-                for field in self.get_form_fields()
-            ]
-
-            # Get all submissions for current page
-            submissions = self.get_submission_class().objects.filter(page=self)
-            for submission in submissions:
-                data = submission.get_data()
-
-                # Count results for each question
-                for name, label in data_fields:
-                    answer = data.get(name)
-                    if answer is None:
-                        # Something wrong with data.
-                        # Probably you have changed questions
-                        # and now we are receiving answers for old questions.
-                        # Just skip them.
-                        continue
-
-                    if type(answer) is list:
-                        # answer is a list if the field type is 'Checkboxes'
-                        answer = u', '.join(answer)
-
-                    question_stats = results.get(label, {})
-                    question_stats[answer] = question_stats.get(answer, 0) + 1
-                    results[label] = question_stats
-
-            context.update({
-                'results': results,
-            })
-
-        return context
 
     def get_data_fields(self):
         data_fields = [
@@ -243,11 +201,10 @@ class MoloSurveyPage(
                         del request.session[session_key_data]
 
                         # Render the landing page
-                        return render(
-                            request,
-                            self.landing_page_template,
-                            self.get_context(request)
-                        )
+                        return redirect(
+                            reverse(
+                                'molo.surveys:success', args=(self.slug, )))
+
             else:
                 # If data for step is invalid
                 # we will need to display form again with errors,
@@ -279,8 +236,15 @@ class MoloSurveyPage(
 
         if request.method == 'POST':
             form = self.get_form(request.POST, page=self, user=request.user)
+
             if form.is_valid():
                 self.set_survey_as_submitted_for_session(request)
+                self.process_form_submission(form)
+
+                # render the landing_page
+                return redirect(
+                    reverse('molo.surveys:success', args=(self.slug, )))
+
         return super(MoloSurveyPage, self).serve(request, *args, **kwargs)