+++ CHANGES.rst
@@ -6,6 +6,14 @@ CHANGE LOG
 - Added merged cms functionality to surveys
 - Only able to see relevant surveys for site in admin and csv
 
+2.3.0
+-----
+- Add a success url after user submit answers to a survey
+
+2.2.2
+-----
+- Create a success page after user submit answers to a survey
+
 2.2.1
 -----
 - Bug Fix: Survey model inherited from non routable page mixin